SRP refactor maintenance
