+++ hapi-fhir-storage-batch2/src/test/java/ca/uhn/fhir/batch2/model/JobWorkCursorTest.java
@@ -1,7 +1,7 @@
 package ca.uhn.fhir.batch2.model;
 
-import ca.uhn.fhir.batch2.impl.BaseBatch2Test;
-import ca.uhn.fhir.batch2.impl.TestJobParameters;
+import ca.uhn.fhir.batch2.coordinator.BaseBatch2Test;
+import ca.uhn.fhir.batch2.coordinator.TestJobParameters;
 import ca.uhn.fhir.rest.server.exceptions.InternalErrorException;
 import org.junit.jupiter.api.BeforeEach;
 import org.junit.jupiter.api.Test;
